refactor(wiki): teach users how to navigate
